@@ -1,28 +1,35 @@
 {
   "name": "uni_crash_log",
+  "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite --host",
     "build": "vite build",
-    "serve": "vite preview"
+    "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^0.27.2",
+    "axios": "^1.5.1",
     "svelte-spa-router": "^3.3.0"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.0.7",
-    "@unocss/core": "^0.45.22",
-    "@unocss/transformer-directives": "^0.45.22",
-    "postcss": "^8.4.16",
-    "postcss-import": "^15.0.0",
+    "@sveltejs/vite-plugin-svelte": "^2.4.6",
+    "@unocss/extractor-svelte": "^0.57.1",
+    "postcss": "^8.4.31",
+    "postcss-import": "^15.1.0",
     "postcss-load-config": "^4.0.1",
-    "postcss-nested": "^5.0.6",
-    "postcss-preset-env": "^7.8.2",
-    "sass": "^1.54.9",
-    "svelte": "^3.50.1",
-    "svelte-preprocess": "^4.10.7",
-    "unocss": "^0.45.22",
-    "vite": "^3.1.3"
+    "postcss-nested": "^6.0.1",
+    "postcss-preset-env": "^9.2.0",
+    "sass": "^1.69.4",
+    "svelte": "^4.2.2",
+    "svelte-preprocess": "^5.0.4",
+    "unocss": "^0.57.1",
+    "vite": "^4.5.0",
+    "@iconify-json/logos": "^1.1.37",
+    "@unocss/core": "^0.57.1",
+    "@unocss/preset-icons": "^0.57.1",
+    "svelte-check": "^3.5.2",
+    "@unocss/postcss": "^0.57.1"
+    
   }
 }
