@@ -1,4 +1,7 @@
-const { postcss } = require('svelte-preprocess');
-module.exports = {
-  preprocess: [postcss()],
-};
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
+
+export default {
+  // Consult https://svelte.dev/docs#compile-time-svelte-preprocess
+  // for more information about preprocessors
+  preprocess: [vitePreprocess()],
+}
