@@ -7,7 +7,7 @@
  */
 import './index.css';
 import App from './App.svelte';
-import 'uno.css'
+import 'virtual:uno.css'
 
 const app = new App({
 	target: document.body,
