@@ -7,20 +7,12 @@
  */
 import { defineConfig } from 'vite'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
-import Unocss from 'unocss/vite'
-import { extractorSvelte } from '@unocss/core'
-import transformerDirective from '@unocss/transformer-directives'
+import UnoCSS from 'unocss/vite'
 
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [
-        Unocss({
-            extractors: [extractorSvelte],
-            transformers: [
-                transformerDirective(),
-            ],
-            /* more options */
-        }),
+        UnoCSS(),
         svelte()
     ],
     server: {
